Basis for AdminHome is done and works
