+++ app/imports/ui/pages/AdminHome.jsx
@@ -1,30 +1,17 @@
 import React from 'react';
-import { Col, Row, Container, Tab, Tabs } from 'react-bootstrap';
-import IngredientsListAdmin from '../components/IngredientsListAdmin';
-import IngredientsAllergiesListAdmin from '../components/IngredientsAllergiesListAdmin';
-import ProfilesListAdmin from '../components/ProfilesListAdmin';
-import VendorsListAdmin from '../components/VendorsListAdmin';
-import RecipesListAdmin from '../components/RecipesListAdmin';
-import RecipesIngredientsListAdmin from '../components/RecipesIngredientsListAdmin';
-import VendorsIngredientsListAdmin from '../components/VendorsIngredientsListAdmin';
+import { Container, Tab, Tabs } from 'react-bootstrap';
+import DataListsAdmin from '../components/DataListsAdmin';
+import RelationListsAdmin from '../components/RelationListsAdmin';
 
-/* Organize all admin data tables */
+/* This component is merely to organize all admin data */
 const AdminHome = () => (
   <Container className="p-0 text-center">
     <Tabs fill>
-      <Tab eventKey="datadb" title="Data Collections">
-        <Row>
-          <Col><IngredientsListAdmin /></Col>
-          <Col><ProfilesListAdmin /><VendorsListAdmin /></Col>
-        </Row>
-        <Row><RecipesListAdmin /></Row>
+      <Tab eventKey="relaDb" title="Relational Collections">
+        <RelationListsAdmin />
       </Tab>
-      <Tab eventKey="reladb" title="Relational Collections">
-        <Col>
-          <Row><IngredientsAllergiesListAdmin /></Row>
-          <Row><RecipesIngredientsListAdmin /></Row>
-          <Row><VendorsIngredientsListAdmin /></Row>
-        </Col>
+      <Tab eventKey="dataDb" title="Data Collections">
+        <DataListsAdmin />
       </Tab>
       <Tab eventKey="search" title="Search Test" />
     </Tabs>
